requirements.txt
flask
line-bot-sdk>=3.0.0
anthropic
python-dotenv
gunicorn
requests
